repo cleanup
moving file manager and dense packer into their own files to declutter RexDB
